test: added unit tests for sayt performance metric functions

+++ src/survey_assist_eval/evaluation/sayt/performance_metrics_functions.py
@@ -194,7 +194,7 @@ def add_sayt_metrics_columns(
 
 
 def summarise_performance_metrics(
-    df, k_values: list[int], ave_time_per_query: int
+    df, k_values: list[int], ave_time_per_query: float
 ) -> SAYTPerformanceMetrics:
     """Summarize performance metrics across the DataFrame.
